iTpes updated jal jalr

+++ simulator/src/bTypes.cpp
@@ -6,10 +6,10 @@ using namespace std;
 uint8_t * beq(line &instr, uint32_t * reg_ptr, uint8_t * prgm_counter){
 	uint32_t *r1,*r2;
 	int32_t imm = 0;
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_12)>>19);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_11)<<4);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_10_5)>>20);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_4_1)>>7);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_12)>>19);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_11)<<4);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_10_5)>>20);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_4_1)>>7);
 
 	r1 = reg_ptr+((instr.instr & EXTRACT_RBS_R1)>>15);
 	r2 = reg_ptr+((instr.instr & EXTRACT_RBS_R2)>>20);
@@ -23,10 +23,10 @@ uint8_t * beq(line &instr, uint32_t * reg_ptr, uint8_t * prgm_counter){
 uint8_t * bne(line &instr, uint32_t * reg_ptr, uint8_t * prgm_counter){
 	uint32_t *r1,*r2;
 	int32_t imm = 0;
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_12)>>19);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_11)<<4);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_10_5)>>20);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_4_1)>>7);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_12)>>19);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_11)<<4);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_10_5)>>20);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_4_1)>>7);
 
 	r1 = reg_ptr+((instr.instr & EXTRACT_RBS_R1)>>15);
 	r2 = reg_ptr+((instr.instr & EXTRACT_RBS_R2)>>20);
@@ -40,10 +40,10 @@ uint8_t * bne(line &instr, uint32_t * reg_ptr, uint8_t * prgm_counter){
 uint8_t * blt(line &instr, uint32_t * reg_ptr, uint8_t * prgm_counter){
 	uint32_t *r1,*r2;
 	int32_t imm = 0;
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_12)>>19);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_11)<<4);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_10_5)>>20);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_4_1)>>7);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_12)>>19);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_11)<<4);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_10_5)>>20);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_4_1)>>7);
 
 	r1 = reg_ptr+((instr.instr & EXTRACT_RBS_R1)>>15);
 	r2 = reg_ptr+((instr.instr & EXTRACT_RBS_R2)>>20);
@@ -57,10 +57,10 @@ uint8_t * blt(line &instr, uint32_t * reg_ptr, uint8_t * prgm_counter){
 uint8_t * bge(line &instr, uint32_t * reg_ptr, uint8_t * prgm_counter){
 	uint32_t *r1,*r2;
 	int32_t imm = 0;
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_12)>>19);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_11)<<4);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_10_5)>>20);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_4_1)>>7);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_12)>>19);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_11)<<4);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_10_5)>>20);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_4_1)>>7);
 
 	r1 = reg_ptr+((instr.instr & EXTRACT_RBS_R1)>>15);
 	r2 = reg_ptr+((instr.instr & EXTRACT_RBS_R2)>>20);
@@ -74,10 +74,10 @@ uint8_t * bge(line &instr, uint32_t * reg_ptr, uint8_t * prgm_counter){
 uint8_t * bltu(line &instr, uint32_t * reg_ptr, uint8_t * prgm_counter){
 	uint32_t *r1,*r2;
 	int32_t imm = 0;
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_12)>>19);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_11)<<4);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_10_5)>>20);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_4_1)>>7);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_12)>>19);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_11)<<4);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_10_5)>>20);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_4_1)>>7);
 
 	r1 = reg_ptr+((instr.instr & EXTRACT_RBS_R1)>>15);
 	r2 = reg_ptr+((instr.instr & EXTRACT_RBS_R2)>>20);
@@ -91,10 +91,10 @@ uint8_t * bltu(line &instr, uint32_t * reg_ptr, uint8_t * prgm_counter){
 uint8_t * bgeu(line &instr, uint32_t * reg_ptr, uint8_t * prgm_counter){
 	uint32_t *r1,*r2;
 	int32_t imm = 0;
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_12)>>19);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_11)<<4);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_10_5)>>20);
-	imm = imm | ((instr.instr & EXTRACT_B_IMM_4_1)>>7);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_12)>>19);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_11)<<4);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_10_5)>>20);
+	imm = imm | (((int32_t)instr.instr & EXTRACT_B_IMM_4_1)>>7);
 
 	r1 = reg_ptr+((instr.instr & EXTRACT_RBS_R1)>>15);
 	r2 = reg_ptr+((instr.instr & EXTRACT_RBS_R2)>>20);
